Minor tweaks from looking at EngineYard

+++ .bashrc
@@ -1,6 +1,12 @@
 # .bashrc
 # Executed for interactive non-login shells
 
+# Test for an interactive shell.
+if [[ $- != *i* ]] ; then
+        # Shell is non-interactive.  Be done now!
+        return
+fi
+
 # Aliases
 alias rm='rm -i'
 alias mv='mv -i'
@@ -11,3 +17,4 @@ if [ "$TERM" != "dumb" ]; then
         eval `dircolors -b`
         alias ls='ls --color=auto'
 fi
+
